Add TaskAddField component.
Text field where end users can input their tasks for items to show up
on the tasks list above. When the end user presses Enter on their
keyboard after entering a string of text, a POST request is sent to the
server, where the database is updated with the relevant name, date,
priorty, and folder of the task.