Add FoldersBar component.
The FoldersBar component holds a list of folderItem components. Each
folderItem has an edit and delete button, along with the name of the
folder. There is always a Main folder which is the end user's default
folder. Deleting folder sends a DELETE request to the server to the
specified folde to be deleted.