Add AddInput and SearchInput components
When the folder bar is hidden, the input field displays the input field
for searching for tasks. On key up, the search query is sent to the
server, and the response contains the items that match the name query.
When not hidden, the field displays the input field for adding new
folders. When the end user adds a new folder, a POST request is sent
from the client to the server to adda new folder with the provided name.