Add TaskList, TaskItem, TaskButtons components
On render, the TaskList component requests the task items that belong to
the current folder (Main being the default). The TaskList component
consists of TaskItem components that represent tasks, along with
functionality like prioritizing, setting due dates, editing text, and
deleting. When the end user clicks one of these functionality buttons,
a request is sent to the server to update the database with new info.